+++ src/game/tests.rs
@@ -596,6 +596,7 @@ mod land_characteristics;
 mod landwalk;
 mod living_artifact;
 mod looking_at_another_library;
+mod m13_more_stale;
 mod m13_stale_audits;
 mod m14_stale_audits;
 mod mana_ability_costs;